// frontend/src/types/StatsResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { VersionStat } from "./VersionStat";

export type StatsResponse = { total_servers: number, cracked_servers: number, online_servers: number, crashed_servers: number, forge_servers: number, spoofable_servers: number, total_players: number, admin_players: number, avg_ping: number | null, version_distribution: Array<VersionStat>, };